moved merge_activities helper function from LPG example to process_model_data.py
